+++ src/components/__tests__/ui/BaseCardCitySkeleton.spec.ts
@@ -11,7 +11,7 @@ describe('BaseCardCitySkeleton', () => {
     );
     expect(wrapper.html()).toContain('relative z-30');
     expect(wrapper.html()).toContain(
-      'w-[60px] h-[60px] bg-gray-300 rounded-[20px] sm:w-[75px] sm:h-[75px]',
+      'w-[60px] h-[60px] bg-gray-300 rounded-[20px] md:w-[75px] md:h-[75px]',
     );
     expect(wrapper.html()).toContain(
       'absolute top-0 left-0 right-0 z-20 w-full h-full bg-gray-600 opacity-50',